# openapi.yaml
openapi: 3.1.0
info:
  title: wa-hub-demo — WhatsApp HTTP API
  version: 0.2.0
  description: >
    Self-hosted WhatsApp HTTP API (Baileys). REST + JSON, with HMAC-signed outbound webhooks.
    PRECONDITION: a WhatsApp device must be paired once (scan a QR at GET /pair) before any send
    works — otherwise send routes return 503 not_connected. Confirm readiness with
    GET /api/instance/status (connection must be "connected"). See docs/INTEGRATION.md.
  license:
    name: MIT
servers:
  - url: https://wa.yourdomain.com
    description: Your stable subdomain (recommended) — or the ephemeral https://<random>.trycloudflare.com tunnel URL
  - url: http://127.0.0.1:3060
    description: On the server itself (loopback)
security:
  - bearerAuth: []
...
paths:
...
  /api/groups:
    get:
      tags: [groups]
      summary: List groups the paired account participates in
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        jid: { type: string }
                        name: { type: string }
                        participants: { type: integer }
                        owner: { type: string, nullable: true }
                        creation: { type: integer, nullable: true }
                        announce: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
    post:
      tags: [groups]
      summary: Create a group
      description: >
        ⚠ Bulk-creating groups or adding strangers is a top ban signal — use only with consenting participants.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject]
              properties:
                subject: { type: string, minLength: 1, maxLength: 100 }
                participants: { type: array, maxItems: 50, default: [], items: { type: string } }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  jid: { type: string }
                  subject: { type: string }
                  participants: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The HUB_TOKEN.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Required on PUT /webhook and POST /logout only when ADMIN_TOKEN is configured.
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid HUB_TOKEN
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BadRequest:
      description: >
        Invalid body / recipient / JSON — e.g. {"error":"invalid_body"} (with Zod issues) or
        {"error":"invalid_json"}. Oversized base64 media (…Base64 over the 20 MB decoded cap)
        also lands here as {"error":"invalid_body"} — "file_too_large" appears only inside
        issues[].message, NOT as the top-level error.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotConnected:
      description: WhatsApp is not paired/connected yet
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }